requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-modeltranslation==0.18.4
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
